// frontend/src/api/models/memoryAgentConfig.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryAgentModelOption } from './memoryAgentModelOption';
import type { MemoryAgentToolsetOption } from './memoryAgentToolsetOption';

export interface MemoryAgentConfig {
  models: MemoryAgentModelOption[];
  toolsets: MemoryAgentToolsetOption[];
}