// frontend/src/api/models/memoryAgentToolsetOption.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryAgentToolsetOptionConfigSchema } from './memoryAgentToolsetOptionConfigSchema';
import type { ToolsetKind } from './toolsetKind';

export interface MemoryAgentToolsetOption {
  configSchema: MemoryAgentToolsetOptionConfigSchema;
  defaultEnabled: boolean;
  description: string;
  excludes: string[];
  kind: ToolsetKind;
  label: string;
  name: string;
  requires: string[];
}